+++ ConsoleTests/CriticalSectionTests.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace ConsoleTests
+{
+    static class CriticalSectionTests
+    {
+
+        public static void Start() 
+        {
+        
+        }
+    }
+}
